requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-facebook==6.0.3
django-filter==2.0.0
djangorestframework==3.9.0
html5lib==1.0.1
idna==2.7
Markdown==3.0.1
numpy==1.16.2
oauthlib==2.1.0
Pillow==5.3.0
pkg-resources==0.0.0
pyBarcode==0.7
PyJWT==1.6.4
pymatlab==0.2.3
PyPDF2==1.26.0
python-barcode==0.9.0
python3-openid==3.1.0
pytz==2018.7
reportlab==3.5.13
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
Unidecode==1.0.22
urllib3==1.24.1
webencodings==0.5.1
xhtml2pdf==0.2.3
